Implement "readability optimization" for integers.
CVC4 tends to use very large numbers in situations then possible to use smaller.
For example, it can produce -2841557 as a value for 'x' in constraint 'x < 0'.
Test cases that include such numbers hard to read.
New optimization pass tries to apply new constraints to integer variablesstep-by-step.
It first trise '-10 < x < 10', then '-100 < x < 100' end so on.
Only constraints which do not damage the PC are appliable.