Implement alpha version of feedback-directed random test generator.
The generator works as a llvm-pass. It gets a file with code of the target function and then generates test suite automatically.
Firstly, it extracts an interface of the target function. Secondly, it produces set of generators for each function's argument.
After that, it gets argument values from the genrators and then runs target function in a JIT-interpreter.
The interpreter is always run in a different (child) process.
If child process terminates normally it produces corresponding result by itself.
But if the child process crashes, the parent (generator) handles status of the crash and print an appropriate message.
The idea is to take successfull runs and then alterate them in order to achive new crashing results.